Hard Fork Prototype 0, initial release.
HFP0 is an early hardfork prototype derived from Classic 0.12.0cl1 and satoshisbitcoin.

It is NOT ready for production use - do not use this client with real money (i.e. your existing Bitcoin wallet).
Please use regtest or testnet to play with this code - avoid testing it on mainnet as it does not separate as cleanly as it should (it bans peers to split off the network).

See README.md for more details.

Fork trigger heights:
- mainnet: 666666
- testnet: 9999999
- regtest: 1000